refactor: just+python → pnpm+node
