+++ .github/workflows/gh-pages.yml
@@ -28,13 +28,18 @@ jobs:
       url: ${{ steps.deployment.outputs.page_url }}
     steps:
       - uses: actions/checkout@v4
-      - uses: taiki-e/install-action@v2
+      - uses: pnpm/action-setup@v4
         with:
-          tool: just
+          version: 10
+      - uses: actions/setup-node@v4
+        with:
+          node-version: 22
+          cache: pnpm
       - uses: typst-community/setup-typst@v4
       - name: Install fonts
         run: sudo apt install fonts-noto-cjk
-      - run: just compile
+      - run: pnpm install
+      - run: pnpm build
       - name: Setup Pages
         uses: actions/configure-pages@v5
       - name: Upload artifact
