--- content/snapshot/example.json
@@ -1,8 +1,0 @@
-{
-  "example-a94a8fe5ccb19ba61c4c0873d391e987982fbbd3": {
-    "output": "dist/assets/example-a94a8fe5ccb19ba61c4c0873d391e987982fbbd3.svg"
-  },
-  "example-ccf587c77d3c946812e21674ed3b95cb47ab0d6d": {
-    "output": "dist/assets/example-ccf587c77d3c946812e21674ed3b95cb47ab0d6d.svg"
-  }
-}
