--- examples/thesis.bib
@@ -1,11 +1,0 @@
-@thesis{王楠2016,
-  title = {在“共产主义视镜”下想象科学 ——“十七年”期间的中国科幻文学与科学话语},
-  author = {王楠},
-  date = {2016-08-05},
-  institution = {新加坡国立大学},
-  location = {新加坡},
-  url = {https://scholarbank.nus.edu.sg/handle/10635/132143},
-  urldate = {2025-02-15},
-  langid = {chinese}
-}
-% 王楠. 在“共产主义视镜”下想象科学 ——“十七年”期间的中国科幻文学与科学话语[D/OL]. 新加坡: 新加坡国立大学, 2016[2025-02-15]. https://scholarbank.nus.edu.sg/handle/10635/132143.
